description edited, handled int vs unsigned warning

+++ src/filltree.cpp
@@ -11,7 +11,7 @@ std::vector<int> children_to_parentC(std::vector< std::vector<int> > children)
   {
     for (unsigned int j = 0; j < children[i].size(); j++)
     {
-      if (children[i][j] > out.size() || children[i][j] < 1)
+      if (children[i][j] > (int)out.size() || children[i][j] < 1)
         Rcpp::stop("index out of bounds, (@children_to_parent)");
       out[children[i][j]-1] = i + 1; // '-1' and '+1' to make it one-based
     }
@@ -30,7 +30,7 @@ std::vector< std::vector<int> > parent_to_childrenC(std::vector<int> parent)
 
   for (unsigned int i = 0; i < parent.size(); i++)
   {
-     if (parent[i] < 0 || parent[i] > out.size())
+     if (parent[i] < 0 || parent[i] > (int)out.size())
        Rcpp::stop("index out of bounds, (@parent_to_children)");
      if (parent[i] >= 1) // index zero, which indicates the root node
        out[parent[i]-1].push_back(i + 1);
